# docs/api/swagger.yaml
basePath: /api
definitions:
...
  models.ClusterSettings:
    properties:
      hosts:
        items:
          $ref: '#/definitions/models.HostConnection'
        type: array
      id:
        type: string
      selected_checks:
        items:
          type: string
        type: array
    type: object
  models.HostConnection:
    properties:
      address:
        type: string
      name:
        type: string
      user:
        type: string
    type: object
...
info:
  contact:
    email: trento-project@suse.com
    name: Trento Project
    url: https://www.trento-project.io
  description: Trento API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Trento API
  version: "1.0"
paths:
...
  /clusters/settings:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ClusterSettings'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Retrieve Settings for all the clusters. Cluster's Selected checks and
        Hosts connection settings
...
schemes:
- http
swagger: "2.0"